requirements.txt
requests==2.31.0
python-telegram-bot==13.15
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.0
